Loaded titles to mysql and implemented a fuzzy search.
Redis seems like a good option especially with redisearch and a built in fuzzy library. Couldnt use it as it is linux only and my debian isnt set up for mysql or node *not for a lack of trying. worth looking for in the future as fuzzy is O(nm) and scales with the size of the query